feat: define routing and add crud for each tables

+++ routes/api.php
@@ -2,21 +2,26 @@
 
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
-use App\Http\Controllers\Api\ProductController;
-use App\Http\Controllers\Api\TransactionController;
+use App\Http\Controllers\Auth\LoginController;
+use App\Http\Controllers\Auth\RegisterController;
+use App\Http\Controllers\ProductController;
+use App\Http\Controllers\TransactionController;
+use App\Http\Middleware\AuthToken;
 
-// Product Routes
-Route::get('/products', [ProductController::class, 'index']);
-Route::post('/products', [ProductController::class, 'store']);
-Route::get('/products/{id}', [ProductController::class, 'show']);
-Route::put('/products/{id}', [ProductController::class, 'update']);
-Route::delete('/products/{id}', [ProductController::class, 'destroy']);
+Route::post('/login', [LoginController::class, 'login'])->withoutMiddleware(AuthToken::class);
+Route::post('/signup', [RegisterController::class, 'register'])->withoutMiddleware(AuthToken::class);
+Route::post('/logout', [LoginController::class, 'logout']);
 
-// Transaction Routes
-Route::get('/transactions', [TransactionController::class, 'index']);
-Route::post('/transactions', [TransactionController::class, 'store']);
-Route::get('/transactions/{id}', [TransactionController::class, 'show']);
-Route::put('/transactions/{id}', [TransactionController::class, 'update']);
-Route::delete('/transactions/{id}', [TransactionController::class, 'destroy']);
+Route::middleware(AuthToken::class)->group(function () {
+    Route::get('/products/{userId}', [ProductController::class, 'index']);
+    Route::post('/products/{userId}', [ProductController::class, 'store']);
+    Route::get('/products/{userId}/{id}', [ProductController::class, 'show']);
+    Route::put('/products/{userId}/{id}', [ProductController::class, 'update']);
+    Route::delete('/products/{userId}/{id}', [ProductController::class, 'destroy']);
 
-Route::get('/products-list', [TransactionController::class, 'getProducts']);
+    Route::get('/transactions/{userId}', [TransactionController::class, 'index']);
+    Route::post('/transactions/{userId}', [TransactionController::class, 'store']);
+    Route::get('/transactions/{userId}/{id}', [TransactionController::class, 'show']);
+    Route::put('/transactions/{userId}/{id}', [TransactionController::class, 'update']);
+    Route::delete('/transactions/{userId}/{id}', [TransactionController::class, 'destroy']);
+});
